Move DfESignIn class to root namespace
The bypass? method will be used for SupportInterface and CheckRecords namespaces.
