+++ app/lib/dfe_sign_in.rb
@@ -1,0 +1,15 @@
+require "hosting_environment"
+
+class DfESignIn
+  def self.bypass?
+    bypass_active? || review_app?
+  end
+
+  def self.bypass_active?
+    HostingEnvironment.test_environment? && ENV["BYPASS_DSI"] == "true"
+  end
+
+  def self.review_app?
+    HostingEnvironment.review?
+  end
+end
